fixing code climate issue
