+++ pkg/gofr/datasource/opentsdb/preprocess.go
@@ -9,43 +9,8 @@ import (
 	"go.opentelemetry.io/otel/trace"
 	"net/http"
 	"strings"
+	"time"
 )
-
-// Annotation holds parameters for querying or managing annotations via the /api/annotation endpoint in OpenTSDB.
-// Used for logging notes on events at specific times, often tied to time series data, mainly for graphing or API queries.
-type Annotation struct {
-	// StartTime is the Unix epoch timestamp (in seconds) for when the event occurred. This is required.
-	StartTime int64 `json:"startTime,omitempty"`
-
-	// EndTime is the optional Unix epoch timestamp (in seconds) for when the event ended, if applicable.
-	EndTime int64 `json:"endTime,omitempty"`
-
-	// TSUID is the optional time series identifier if the annotation is linked to a specific time series.
-	TSUID string `json:"tsuid,omitempty"`
-
-	// Description is a brief, optional summary of the event (recommended to keep under 25 characters for display purposes).
-	Description string `json:"description,omitempty"`
-
-	// Notes is an optional, detailed description of the event.
-	Notes string `json:"notes,omitempty"`
-
-	// Custom is an optional key/value map to store any additional fields and their values.
-	Custom map[string]string `json:"custom,omitempty"`
-}
-
-// AnnotationResponse encapsulates the response data and status when interacting with the /api/annotation endpoint.
-type AnnotationResponse struct {
-
-	// Annotation holds the associated annotation object.
-	Annotation
-
-	// ErrorInfo contains details about any errors that occurred during the request.
-	ErrorInfo map[string]any `json:"error,omitempty"`
-
-	logger Logger
-	tracer trace.Tracer
-	ctx    context.Context
-}
 
 // QueryParam is the structure used to hold the querying parameters when calling /api/query.
 // Each attributes in QueryParam matches the definition in
@@ -287,64 +252,6 @@ func isEmptyPutParam(param string) bool {
 	return strings.TrimSpace(param) == ""
 }
 
-// QueryResponse acts as the implementation of Response in the /api/query scene.
-// It holds the status code and the response values defined in the
-// [OpenTSDB Official Docs]: http://opentsdb.net/docs/build/html/api_http/query/index.html.
-type QueryResponse struct {
-	QueryRespCnts []QueryRespItem `json:"queryRespCnts"`
-	ErrorMsg      map[string]any  `json:"error"`
-	logger        Logger
-	tracer        trace.Tracer
-	ctx           context.Context
-}
-
-// QueryRespItem acts as the implementation of Response in the /api/query scene.
-// It holds the response item defined in the
-// [OpenTSDB Official Docs]: http://opentsdb.net/docs/build/html/api_http/query/index.html.
-type QueryRespItem struct {
-	// Name of the metric retrieved for the time series
-	Metric string `json:"metric"`
-
-	// A list of tags only returned when the results are for a single time series.
-	// If results are aggregated, this value may be null or an empty map
-	Tags map[string]string `json:"tags"`
-
-	// If more than one time series were included in the result set, i.e. they were aggregated,
-	// this will display a list of tag names that were found in common across all time series.
-	// Note that: Api Doc uses 'aggregatedTags', but actual response uses 'aggregateTags'
-	AggregatedTags []string `json:"aggregateTags"`
-
-	// Retrieved data points after being processed by the aggregators. Each data point consists
-	// of a timestamp and a value, the format determined by the serializer.
-	// For the JSON serializer, the timestamp will always be a Unix epoch style integer followed
-	// by the value as an integer or a floating point.
-	// For example, the default output is "dps"{"<timestamp>":<value>}.
-	// By default, the timestamps will be in seconds. If the msResolution flag is set, then the
-	// timestamps will be in milliseconds.
-	//
-	// Because the elements of map is out of order, using common way to iterate Dps will not get
-	// data points with timestamps out of order.
-	// So be aware that one should use '(qri *QueryRespItem) GetDataPoints() []*DataPoint' to
-	// acquire the real ascending data points.
-	Dps map[string]any `json:"dps"`
-
-	// If the query retrieved annotations for time series over the requested timespan, they will
-	// be returned in this group. Annotations for every time series will be merged into one set
-	// and sorted by start_time. Aggregator functions do not affect annotations, all annotations
-	// will be returned for the span.
-	// The value is optional.
-	Annotations []Annotation `json:"annotations,omitempty"`
-
-	// If requested by the user, the query will scan for global annotations during
-	// the timespan and the results returned in this group.
-	// The value is optional.
-	GlobalAnnotations []Annotation `json:"globalAnnotations,omitempty"`
-
-	logger Logger
-	tracer trace.Tracer
-	ctx    context.Context
-}
-
 // QueryLastParam is the structure used to hold
 // the querying parameters when calling /api/query/last.
 // Each attributes in QueryLastParam matches the definition in
@@ -380,40 +287,6 @@ type SubQueryLast struct {
 	// Note that if no tags are specified, all metrics in the system will be aggregated into the results.
 	// It will be deprecated in OpenTSDB 2.2.
 	Tags map[string]string `json:"tags,omitempty"`
-}
-
-// QueryLastResponse acts as the implementation of Response in the /api/query/last scene.
-// It holds the status code and the response values defined in the
-// [OpenTSDB Official Docs]: http://opentsdb.net/docs/build/html/api_http/query/last.html.
-type QueryLastResponse struct {
-	QueryRespCnts []QueryRespLastItem `json:"queryRespCnts,omitempty"`
-	ErrorMsg      map[string]any      `json:"error"`
-	logger        Logger
-	tracer        trace.Tracer
-	ctx           context.Context
-}
-
-// QueryRespLastItem acts as the implementation of Response in the /api/query/last scene.
-// It holds the response item defined in the
-// [OpenTSDB Official Docs]: http://opentsdb.net/docs/build/html/api_http/query/last.html.
-type QueryRespLastItem struct {
-	// Name of the metric retreived for the time series.
-	// Only returned if resolve was set to true.
-	Metric string `json:"metric"`
-
-	// A list of tags only returned when the results are for a single time series.
-	// If results are aggregated, this value may be null or an empty map.
-	// Only returned if resolve was set to true.
-	Tags map[string]string `json:"tags"`
-
-	// A Unix epoch timestamp, in milliseconds, when the data point was written.
-	Timestamp int64 `json:"timestamp"`
-
-	// The value of the data point enclosed in quotation marks as a string
-	Value string `json:"value"`
-
-	// The hexadecimal TSUID for the time series
-	TSUID string `json:"tsuid"`
 }
 
 func getQueryBodyContents(param any) (string, error) {
@@ -495,4 +368,53 @@ func isValidQueryLastParam(param *QueryLastParam) bool {
 	}
 
 	return true
+}
+
+func (c *Client) operateAnnotation(ctx context.Context, queryAnnotation any, resp any, method, operation string) error {
+	span := c.addTrace(ctx, operation)
+
+	status := StatusFailed
+
+	var message string
+
+	defer sendOperationStats(c.logger, time.Now(), operation, &status, &message, span)
+
+	annotation, ok := queryAnnotation.(*Annotation)
+	if !ok {
+		return errors.New("invalid annotation type. Must be *Annotation")
+	}
+
+	annResp, ok := resp.(*AnnotationResponse)
+	if !ok {
+		return errors.New("invalid response type. Must be *AnnotationResponse")
+	}
+
+	if !c.isValidOperateMethod(ctx, method) {
+		message = fmt.Sprintf("invalid annotation operation method: %s", method)
+		return errors.New(message)
+	}
+
+	annoEndpoint := fmt.Sprintf("%s%s", c.endpoint, AnnotationPath)
+
+	resultBytes, err := json.Marshal(annotation)
+	if err != nil {
+		message = fmt.Sprintf("marshal annotation response error: %s", err)
+		return errors.New(message)
+	}
+
+	annResp.logger = c.logger
+	annResp.tracer = c.tracer
+	annResp.ctx = ctx
+
+	if err = c.sendRequest(ctx, method, annoEndpoint, string(resultBytes), annResp); err != nil {
+		message = fmt.Sprintf("%s: error while processing %s annotation request to url %q: %s", operation, method, annoEndpoint, err.Error())
+		return err
+	}
+
+	status = StatusSuccess
+	message = fmt.Sprintf("%s: %s annotation request to url %q processed successfully", operation, method, annoEndpoint)
+
+	c.logger.Log("%s request successful", operation)
+
+	return nil
 }
